% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-studies.R
\name{list_studies}
\alias{list_studies}
\title{List available Rapid7 Cybersecurity Studies}
\usage{
list_studies(rapid7_opendata_api_key = rapid7_api_key())
}
\arguments{
\item{rapid7_opendata_api_key}{Your Rapid7 Open Data API key. The various
API interface function look for this key in \code{RAPID7_OPENDATA_API_KEY}.
You can manually provide this key and if the function is used interactively
and the key is not found, you will be prompted to enter the key.}
}
\value{
data frame
}
\description{
List available Rapid7 Cybersecurity Studies
}
\examples{
try(
  list_studies(), silent=TRUE
) -> studies
}
