% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropendata-package.R
\docType{package}
\name{ropendata}
\alias{ropendata}
\alias{ropendata-package}
\title{Query and Download 'Rapid7' Cybersecurity Data Sets}
\description{
'Rapid7' collects cybersecurity data and makes it available via
their 'Open Data' <opendata.rapid7.com> portal which has an 'API'. Tools are
provided to assist in querying for available data sets and downloading any
data set authorized to a registered account.
}
\author{
Bob Rudis (bob_rudis@rapid7.com)
}
