% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropendata-package.R
\docType{package}
\name{ropendata}
\alias{ropendata}
\alias{ropendata-package}
\title{Query and Download 'Rapid7' Cybersecurity Data Sets}
\description{
'Rapid7' collects cybersecurity data and makes it available via
their 'Open Data' <opendata.rapid7.com> portal which has an 'API'. Tools are
provided to assist in querying for available data sets and downloading any
data set authorized to a registered account.
}
\details{
You will need to request a free account on Open Data via
\url{https://opendata.rapid7.com/#register} and then navigate to the "Open Data API"
link there to create both an organizational key and a user key. You can only
use \strong{user keys} with the Open Data API and you will receive error messages
indicating so if you try to use an organizational key.
}
\author{
Bob Rudis (bob_rudis@rapid7.com)
}
\keyword{internal}
